+++ yash-env/src/subshell.rs
@@ -648,8 +648,11 @@ mod tests {
 
             let state = state.borrow();
             let parent_process = &state.processes[&parent_env.main_pid];
-            assert!(!parent_process.blocked_signals().contains(&SIGINT));
-            assert!(!parent_process.blocked_signals().contains(&SIGQUIT));
+            assert_eq!(parent_process.blocked_signals().contains(SIGINT), Ok(false));
+            assert_eq!(
+                parent_process.blocked_signals().contains(SIGQUIT),
+                Ok(false)
+            );
             let child_process = &state.processes[&child_pid];
             assert_eq!(child_process.disposition(SIGINT), Disposition::Ignore);
             assert_eq!(child_process.disposition(SIGQUIT), Disposition::Ignore);